// src/protocols/mrp/generated/messages/audio/GetVolumeMessage.ts
export interface GetVolumeMessage {
  outputDeviceUID?: string | undefined;
  _unknownFields?: { [key: number]: Uint8Array[] } | undefined;
}

function createBaseGetVolumeMessage(): GetVolumeMessage {
  return { outputDeviceUID: "", _unknownFields: {} };
}

export const GetVolumeMessage: MessageFns<GetVolumeMessage> = {
  encode(
    message: GetVolumeMessage,
    writer: BinaryWriter = new BinaryWriter()
  ): BinaryWriter {
    if (
      message.outputDeviceUID !== undefined &&
      message.outputDeviceUID !== ""
    ) {
      writer.uint32(10).string(message.outputDeviceUID);
    }
    if (message._unknownFields !== undefined) {
      for (const [key, values] of globalThis.Object.entries(
        message._unknownFields
      )) {
        const tag = parseInt(key, 10);
        for (const value of values) {
          writer.uint32(tag).raw(value);
        }
      }
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetVolumeMessage {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetVolumeMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.outputDeviceUID = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      const buf = reader.skip(tag & 7);

      const list = message._unknownFields![tag];

      if (list === undefined) {
        message._unknownFields![tag] = [buf];
      } else {
        list.push(buf);
      }
    }
    return message;
  },

  fromJSON(object: any): GetVolumeMessage {
    return {
      outputDeviceUID: isSet(object.outputDeviceUID)
        ? globalThis.String(object.outputDeviceUID)
        : "",
    };
  },

  toJSON(message: GetVolumeMessage): unknown {
    const obj: any = {};
    if (
      message.outputDeviceUID !== undefined &&
      message.outputDeviceUID !== ""
    ) {
      obj.outputDeviceUID = message.outputDeviceUID;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetVolumeMessage>, I>>(
    base?: I
  ): GetVolumeMessage {
    return GetVolumeMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetVolumeMessage>, I>>(
    object: I
  ): GetVolumeMessage {
    const message = createBaseGetVolumeMessage();
    message.outputDeviceUID = object.outputDeviceUID ?? "";
    return message;
  },
};
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}